+++ modules/planning/open_space/dual_variable_warm_start_ipopt_interface.cc
@@ -36,7 +36,7 @@ constexpr std::size_t N = 80;
 
 DualVariableWarmStartIPOPTInterface::DualVariableWarmStartIPOPTInterface(
     int num_of_variables, int num_of_constraints, std::size_t horizon, float ts,
-    const Eigen::MatrixXd& ego, const Eigen::MatrixXd& obstacles_edges_num,
+    const Eigen::MatrixXd& ego, const Eigen::MatrixXi& obstacles_edges_num,
     const std::size_t obstacles_num, const Eigen::MatrixXd& obstacles_A,
     const Eigen::MatrixXd& obstacles_b, const Eigen::MatrixXd& xWS,
     const PlannerOpenSpaceConfig& planner_open_space_config)
