+++ modules/planning/open_space/dual_variable_warm_start_problem.h
@@ -39,7 +39,7 @@ class DualVariableWarmStartProblem {
 
   bool Solve(const std::size_t& horizon, const float& ts,
              const Eigen::MatrixXd& ego, const std::size_t obstacles_num,
-             const Eigen::MatrixXd& obstacles_edges_num,
+             const Eigen::MatrixXi& obstacles_edges_num,
              const Eigen::MatrixXd& obstacles_A,
              const Eigen::MatrixXd& obstacles_b, const Eigen::MatrixXd& xWS,
              Eigen::MatrixXd* l_warm_up, Eigen::MatrixXd* n_warm_up);
